Implement more nodes and add tests

+++ nodes/src/test/java/de/eldoria/bloodnight/nodes/transform/impl/math/FloorNodeTest.java
@@ -3,7 +3,7 @@ package de.eldoria.bloodnight.nodes.transform.impl.math;
 import de.eldoria.bloodnight.nodes.DataType;
 import de.eldoria.bloodnight.nodes.Fields;
 import de.eldoria.bloodnight.nodes.NodeContainer;
-import de.eldoria.bloodnight.nodes.base.Field;
+import de.eldoria.bloodnight.nodes.base.io.Field;
 import de.eldoria.bloodnight.nodes.input.impl.NumberNode;
 import org.junit.jupiter.api.Test;
 
@@ -18,7 +18,7 @@ class FloorNodeTest {
         var floorNode = new FloorNode();
         floorNode.input().set(nodeContainer, Fields.VALUE, new Field(1, Fields.VALUE));
         var output = floorNode.output(nodeContainer);
-        assertEquals(30,  output.get(Fields.RESULT));
+        assertEquals(30, (int) output.get(Fields.RESULT));
         assertEquals(DataType.INTEGER, output.getType(Fields.RESULT));
     }
 }
